reliably terminate on context cancellation
